moved api.wx.util.icon into api.image
