--- amulet_map_editor/api/wx/util/icon.py
@@ -1,19 +1,0 @@
-from amulet_map_editor.api import image
-
-tablericons = image.icon.tablericons
-
-ADD_ICON = tablericons.plus_green
-SUBTRACT_ICON = tablericons.minus_red
-EDIT_ICON = tablericons.edit
-REFRESH_ICON = tablericons.refresh
-
-UP_ARROW = tablericons.chevron_up
-DOWN_ARROW = tablericons.chevron_down
-
-UP_CARET = tablericons.caret_up
-DOWN_CARET = tablericons.caret_down
-
-COLOUR_PICKER = tablericons.color_picker
-TRASH = tablericons.trash
-MAXIMIZE = tablericons.arrows_maximize
-MINIMIZE = tablericons.arrows_minimize
