+++ amulet_map_editor/api/wx/util/key_config.py
@@ -6,7 +6,7 @@ from amulet_map_editor.api.wx.ui.simple import (
 )
 from typing import Dict, Tuple, Optional, Union, Sequence
 
-from amulet_map_editor.api.wx.util.icon import ADD_ICON, SUBTRACT_ICON, EDIT_ICON
+from amulet_map_editor.api.image import ADD_ICON, SUBTRACT_ICON, EDIT_ICON
 
 ModifierKeyType = str
 KeyType = Union[int, str]
